Add more alpha equivalence to Proof_equality
We used to not have any alpha equality, but this is to heavy-handed. We did not
have it because tactics can refer to some binders, such as 'intros until x'.
In practice, the only binders that can be referred to are in the spine of a
goal.
Therefore, we only test the names of the binders in the spine.