Fix the failing tests for proof equality.
The fact that these fail actually indicates that the relaxation performed a few
commits ago is incorrect. We stopped checking for equality of binders in
hypotheses, but these binder names are clearly obversable by applying the
hypothesis, and then running intro. This introduces a new hypothesis, and its
name is chosen based on the name of the original binder...