Adapt to coq/coq#14779 w.r.t. nested lemmas
The changes in this PR are mostly irrelevant to us. The only critical change is
that for commands that are classified to change the proof, it is now checked
that an actual proof must be active (this was not checked before).

Because we classify tactics as having side-effects, they are being re-run at
Qed-time. Normally this is fine (we just make the extra tactic a no-op).
However, when a nested proof exists within the current proof, this somehow
causes there to be no open proof anymore at Qed-time. This is a problem because
after the PR, an error is now thrown.

The solution is to use a CUSTOM classifier that resolves to `vtdefault` at
Qed-time and `vtmodifyproof` otherwise.