Update w.r.t. coq/coq#16503: Do not inline Let's during discharging
Let's can be defined in proof mode and can have huge terms associated with them.
Inlining them into proof states is rather expensive (and counterproductive). So
we just keep them as let's in the proof state. An example of undesired behavior
can be found in ftc/Rolle.v in CoRN.