Adapt to rocq-prover/rocq#21189.
We use sets based on user names rather than canonical names. This may result
in different behaviour in presence of module aliasing, but it is not clear
to me from my weak understanding of the code alone that this is actually not
the desired behaviour.

The new code is backwards compatible in terms of API.